+++ demo/lib/App.tsx
@@ -1,17 +1,26 @@
 import * as React from "react";
 import {
+  Box,
   Button,
-  Checkbox,
-  Container,
+  Drawer,
   Divider,
-  Dropdown,
-  Grid,
-  Icon,
-  Image,
-  Input,
-  Menu,
-  Sidebar,
-} from "semantic-ui-react";
+  FormControlLabel,
+  IconButton,
+  List,
+  ListItem,
+  MenuItem,
+  Select,
+  Slider,
+  Switch,
+  TextField,
+  Typography,
+  SelectChangeEvent,
+} from "@mui/material";
+import ChevronLeftIcon from "@mui/icons-material/ChevronLeft";
+import SearchIcon from "@mui/icons-material/Search";
+import GitHubIcon from "@mui/icons-material/GitHub";
+import ArticleIcon from "@mui/icons-material/Article";
+import EditIcon from "@mui/icons-material/Edit";
 import seqparse from "seqparse";
 
 import Circular from "../../src/Circular/Circular";
@@ -104,12 +113,12 @@ export default class App extends React.Component<any, AppState> {
     viewer: "both",
     zoom: 50,
   };
-  linearRef: React.RefObject<HTMLDivElement> = React.createRef();
-  circularRef: React.RefObject<HTMLDivElement> = React.createRef();
+
+  linearRef: React.RefObject<HTMLDivElement | null> = React.createRef();
+  circularRef: React.RefObject<HTMLDivElement | null> = React.createRef();
 
   componentDidMount = async () => {
     const seq = await seqparse(file);
-
     this.setState({ annotations: seq.annotations, name: seq.name, seq: seq.seq });
   };
 
@@ -138,9 +147,9 @@ export default class App extends React.Component<any, AppState> {
   };
 
   render() {
-    let customChildren = null;
+    let customChildren: any = undefined;
     if (this.state.customChildren) {
-      customChildren = ({ circularProps, linearProps, ...props }) => {
+      customChildren = ({ circularProps, linearProps, ...props }: any) => {
         if (this.state.viewer === "linear") {
           return (
             <div ref={this.linearRef} style={{ height: "100%", width: "100%" }}>
@@ -191,253 +200,364 @@ export default class App extends React.Component<any, AppState> {
     }
 
     return (
-      <div style={{ height: "100vh" }}>
-        <Sidebar.Pushable className="sidebar-container">
-          <Sidebar
-            animation="overlay"
-            as={Menu}
-            id="options-sidebar"
-            stylename="sidebar-container"
-            vertical
-            visible={this.state.showSidebar}
-            onHide={this.handleHide}
-          >
-            <SidebarHeader toggleSidebar={this.toggleSidebar} />
-            <Menu.Item as="a">
+      <Box sx={{ height: "100vh", display: "flex" }}>
+        <Drawer
+          anchor="left"
+          open={this.state.showSidebar}
+          onClose={this.handleHide}
+          sx={{
+            width: 320,
+            flexShrink: 0,
+            "& .MuiDrawer-paper": {
+              width: 320,
+              boxSizing: "border-box",
+            },
+          }}
+        >
+          <SidebarHeader toggleSidebar={this.toggleSidebar} />
+          <List sx={{ px: 2 }}>
+            <ListItem sx={{ px: 0, py: 1, mb: 0 }}>
               <ViewerTypeInput
                 setType={(viewer: string) => {
                   this.setState({ viewer });
                 }}
               />
-            </Menu.Item>
-            <Menu.Item as="a">
+            </ListItem>
+            <Divider sx={{ my: 0 }} />
+            
+            <ListItem sx={{ px: 0, py: 1, mb: 0 }}>
               <LinearZoomInput setZoom={zoom => this.setState({ zoom })} />
-            </Menu.Item>
-            <Menu.Item as="a">
+            </ListItem>
+            <Divider sx={{ my: 0 }} />
+            
+            <ListItem sx={{ px: 0, py: 1, mb: 0 }}>
               <SearchQueryInput setQuery={query => this.setState({ search: { query } })} />
-            </Menu.Item>
-            <Menu.Item as="a" className="options-checkbox">
+            </ListItem>
+            <Divider sx={{ my: 0 }} />
+            
+            <ListItem sx={{ px: 0, py: 1, mb: 0 }}>
               <CheckboxInput
                 label="Show complement"
                 name="showComplement"
+                checked={this.state.showComplement}
                 set={(showComplement: boolean) => this.setState({ showComplement })}
               />
-            </Menu.Item>
-            <Menu.Item as="a" className="options-checkbox">
-              <CheckboxInput label="Show index" name="index" set={showIndex => this.setState({ showIndex })} />
-            </Menu.Item>
-            <Menu.Item as="a" className="options-checkbox">
+            </ListItem>
+            <Divider sx={{ my: 0 }} />
+            
+            <ListItem sx={{ px: 0, py: 1, mb: 0 }}>
+              <CheckboxInput
+                label="Show index"
+                name="index"
+                checked={this.state.showIndex}
+                set={showIndex => this.setState({ showIndex })}
+              />
+            </ListItem>
+            <Divider sx={{ my: 0 }} />
+            
+            <ListItem sx={{ px: 0, py: 1, mb: 0 }}>
               <CheckboxInput
                 label="Custom Children"
                 name="customChildren"
+                checked={this.state.customChildren}
                 set={customChildren => this.setState({ customChildren })}
               />
-            </Menu.Item>
-            <Menu.Item as="a">
+            </ListItem>
+            <Divider sx={{ my: 0 }} />
+            
+            <ListItem sx={{ px: 0, py: 1, mb: 0 }}>
               <EnzymeInput enzymes={this.state.enzymes} toggleEnzyme={this.toggleEnzyme} />
-            </Menu.Item>
-            <SidebarFooter />
-          </Sidebar>
-          <Sidebar.Pusher as={Container} dimmed={this.state.showSidebar} fluid>
-            <div id="seqviz-container">
-              <Header
+            </ListItem>
+            <Divider sx={{ my: 0 }} />
+          </List>
+          <SidebarFooter />
+        </Drawer>
+
+        <Box
+          component="main"
+          sx={{
+            flexGrow: 1,
+            height: "100vh",
+            display: "flex",
+            flexDirection: "column",
+            position: "relative",
+          }}
+        >
+          <Header
+            selection={this.state.selection}
+            showSelectionMeta={this.state.showSelectionMeta}
+            toggleShowSelectionMeta={this.toggleShowSelectionMeta}
+            toggleSidebar={this.toggleSidebar}
+          />
+          <Box
+            id="seqviewer"
+            sx={{
+              height: "calc(100% - 64px)",
+              width: "100%",
+              flex: 1,
+            }}
+          >
+            {this.state.seq && (
+              <SeqViz
+                key={`${this.state.viewer}${this.state.customChildren}`}
+                annotations={this.state.annotations}
+                primers={this.state.primers}
+                enzymes={this.state.enzymes}
+                highlights={[{ start: 0, end: 10 }]}
+                name={this.state.name}
+                onSelection={selection => this.setState({ selection })}
+                    refs={{ circular: this.circularRef as any, linear: this.linearRef as any }}
+                search={this.state.search}
                 selection={this.state.selection}
-                showSelectionMeta={this.state.showSelectionMeta}
-                toggleShowSelectionMeta={this.toggleShowSelectionMeta}
-                toggleSidebar={this.toggleSidebar}
-              />
-              <div id="seqviewer">
-                {this.state.seq && (
-                  <SeqViz
-                    // accession="MN623123"
-                    key={`${this.state.viewer}${this.state.customChildren}`}
-                    annotations={this.state.annotations}
-                    primers={this.state.primers}
-                    enzymes={this.state.enzymes}
-                    highlights={[{ start: 0, end: 10 }]}
-                    name={this.state.name}
-                    onSelection={selection => this.setState({ selection })}
-                    refs={{ circular: this.circularRef, linear: this.linearRef }}
-                    search={this.state.search}
-                    selection={this.state.selection}
-                    seq={this.state.seq}
-                    showComplement={this.state.showComplement}
-                    showIndex={this.state.showIndex}
-                    translations={this.state.translations}
-                    viewer={this.state.viewer as "linear" | "circular"}
-                    zoom={{ linear: this.state.zoom }}
-                  >
-                    {customChildren}
-                  </SeqViz>
-                )}
-              </div>
-            </div>
-          </Sidebar.Pusher>
-        </Sidebar.Pushable>
-      </div>
+                seq={this.state.seq}
+                showComplement={this.state.showComplement}
+                showIndex={this.state.showIndex}
+                translations={this.state.translations}
+                viewer={this.state.viewer as "linear" | "circular"}
+                zoom={{ linear: this.state.zoom }}
+              >
+                {customChildren}
+              </SeqViz>
+            )}
+          </Box>
+        </Box>
+      </Box>
     );
   }
 }
 
 const ViewerTypeInput = ({ setType }: { setType: (viewType: string) => void }) => (
-  <div className="option" id="topology">
-    <span>Topology</span>
-    <Dropdown
+  <Box sx={{ width: "100%", display: "flex", alignItems: "center" }}>
+    <Typography variant="caption" sx={{ mr: 1, minWidth: "80px", fontSize: "0.85rem" }}>
+      Topology
+    </Typography>
+    <Select
       defaultValue="both"
-      fluid
-      options={viewerTypeOptions}
-      selection
-      onChange={(_, data) => {
-        setType(`${data.value}`);
+      fullWidth
+      size="small"
+      onChange={(e: SelectChangeEvent) => {
+        setType(e.target.value);
       }}
-    />
-  </div>
+    >
+      {viewerTypeOptions.map(option => (
+        <MenuItem key={option.key} value={option.value}>
+          {option.text}
+        </MenuItem>
+      ))}
+    </Select>
+  </Box>
 );
 
 const LinearZoomInput = ({ setZoom }: { setZoom: (zoom: number) => void }) => (
-  <div className="option" id="zoom">
-    <span>Zoom</span>
-    <input
-      className="slider"
-      defaultValue="50"
-      id="zoom"
-      max="100"
-      min="1"
-      type="range"
-      onChange={e => {
-        setZoom(parseInt(e.target.value));
+  <Box sx={{ width: "100%", display: "flex", alignItems: "center" }}>
+    <Typography variant="caption" sx={{ mr: 1, minWidth: "80px", fontSize: "0.85rem" }}>
+      Zoom
+    </Typography>
+    <Slider
+      defaultValue={50}
+      min={1}
+      max={100}
+      onChange={(_, value) => {
+        setZoom(value as number);
       }}
+      valueLabelDisplay="auto"
+      sx={{ flex: 1 }}
     />
-  </div>
+  </Box>
 );
 
 const SearchQueryInput = ({ setQuery }: { setQuery: (query: string) => void }) => (
-  <div className="option" id="options-search">
-    <Input icon="search" placeholder="Search..." onChange={(_, data) => setQuery(data.value)} />
-  </div>
+  <Box sx={{ width: "100%", display: "flex", alignItems: "center" }}>
+    <TextField
+      fullWidth
+      size="small"
+      placeholder="Search..."
+      onChange={e => setQuery(e.target.value)}
+      InputProps={{
+        startAdornment: <SearchIcon sx={{ mr: 1, color: "action.active" }} />,
+      }}
+    />
+  </Box>
 );
 
-const CheckboxInput = ({ label, name, set }: { label: string; name: string; set: (v: any) => void }) => (
-  <Checkbox defaultChecked label={label} name={name} toggle onChange={(_, data) => set(data.checked)} />
+const CheckboxInput = ({
+  label,
+  name,
+  checked,
+  set,
+}: {
+  label: string;
+  name: string;
+  checked: boolean;
+  set: (v: any) => void;
+}) => (
+  <FormControlLabel
+    control={<Switch checked={checked} name={name} onChange={(_, checked) => set(checked)} />}
+    label={label}
+    sx={{ 
+      width: "100%", 
+      ml: 0,
+      "& .MuiFormControlLabel-label": {
+        fontSize: "0.85rem"
+      }
+    }}
+  />
 );
 
 const EnzymeInput = ({ enzymes, toggleEnzyme }: { enzymes: string[]; toggleEnzyme: (e: string) => void }) => (
-  <div className="option" id="enzymes">
-    <span>Enzymes</span>
-    <Grid columns={2} id="enzyme-grid">
-      <Grid.Row className="enzyme-grid-row">
-        <Grid.Column className="enzyme-grid-column">
-          <Button
-            active={enzymes.includes("PstI")}
-            className="enzyme-button"
-            color={enzymes.includes("PstI") ? "blue" : null}
-            fluid
-            onClick={() => toggleEnzyme("PstI")}
-          >
-            PstI
-          </Button>
-        </Grid.Column>
-        <Grid.Column className="enzyme-grid-column">
-          <Button
-            active={enzymes.includes("EcoRI")}
-            className="enzyme-button"
-            color={enzymes.includes("EcoRI") ? "blue" : null}
-            fluid
-            onClick={() => toggleEnzyme("EcoRI")}
-          >
-            EcoRI
-          </Button>
-        </Grid.Column>
-      </Grid.Row>
-      <Grid.Row className="enzyme-grid-row">
-        <Grid.Column className="enzyme-grid-column">
-          <Button
-            active={enzymes.includes("XbaI")}
-            className="enzyme-button"
-            color={enzymes.includes("XbaI") ? "blue" : null}
-            fluid
-            onClick={() => toggleEnzyme("XbaI")}
-          >
-            XbaI
-          </Button>
-        </Grid.Column>
-        <Grid.Column className="enzyme-grid-column">
-          <Button
-            active={enzymes.includes("SpeI")}
-            className="enzyme-button"
-            color={enzymes.includes("SpeI") ? "blue" : null}
-            fluid
-            onClick={() => toggleEnzyme("SpeI")}
-          >
-            SpeI
-          </Button>
-        </Grid.Column>
-      </Grid.Row>
-    </Grid>
-  </div>
+  <Box sx={{ width: "100%", display: "flex", alignItems: "center" }}>
+    <Typography variant="caption" sx={{ mr: 1, minWidth: "80px", fontSize: "0.85rem" }}>
+      Enzymes
+    </Typography>
+    <Box sx={{ flex: 1, display: "grid", gridTemplateColumns: "1fr 1fr", gap: 1 }}>
+      <Button
+        fullWidth
+        variant={enzymes.includes("PstI") ? "contained" : "outlined"}
+        onClick={() => toggleEnzyme("PstI")}
+        sx={{
+          textTransform: "none",
+          py: 0.75,
+          fontSize: "0.85rem",
+          color: enzymes.includes("PstI") ? "white" : "primary",
+        }}
+      >
+        PstI
+      </Button>
+      <Button
+        fullWidth
+        variant={enzymes.includes("EcoRI") ? "contained" : "outlined"}
+        onClick={() => toggleEnzyme("EcoRI")}
+        sx={{
+          textTransform: "none",
+          py: 0.75,
+          fontSize: "0.85rem",
+          color: enzymes.includes("EcoRI") ? "white" : "primary",
+        }}
+      >
+        EcoRI
+      </Button>
+      <Button
+        fullWidth
+        variant={enzymes.includes("XbaI") ? "contained" : "outlined"}
+        onClick={() => toggleEnzyme("XbaI")}
+        sx={{
+          textTransform: "none",
+          py: 0.75,
+          fontSize: "0.85rem",
+          color: enzymes.includes("XbaI") ? "white" : "primary",
+        }}
+      >
+        XbaI
+      </Button>
+      <Button
+        fullWidth
+        variant={enzymes.includes("SpeI") ? "contained" : "outlined"}
+        onClick={() => toggleEnzyme("SpeI")}
+        sx={{
+          textTransform: "none",
+          py: 0.75,
+          fontSize: "0.85rem",
+          color: enzymes.includes("SpeI") ? "white" : "primary",
+        }}
+      >
+        SpeI
+      </Button>
+    </Box>
+  </Box>
 );
 
 const SidebarHeader = ({ toggleSidebar }: { toggleSidebar: () => void }) => (
-  <div className="sidebar-header">
-    <div id="header-left">
-      <Image id="seqviz-graphic" src="https://tools.latticeautomation.com/seqviz/seqviz-logo.png" />
-      <h3>Settings</h3>
-    </div>
-    <Button
-      circular
-      className="circular-button"
-      floated="right"
-      icon="angle left"
-      id="sidebar-toggle-close"
-      onClick={toggleSidebar}
-    />
-  </div>
+  <Box
+    sx={{
+      height: "64px",
+      p: 2,
+      boxShadow: "0 2px 8px #f0f1f2",
+      display: "flex",
+      justifyContent: "space-between",
+      alignItems: "center",
+    }}
+  >
+    <Box sx={{ display: "flex", alignItems: "center" }}>
+      <img
+        src="https://tools.latticeautomation.com/seqviz/seqviz-logo.png"
+        alt="SeqViz"
+        style={{ height: "36px", marginRight: "16px" }}
+      />
+      <Typography variant="h6">Settings</Typography>
+    </Box>
+    <IconButton onClick={toggleSidebar}>
+      <ChevronLeftIcon />
+    </IconButton>
+  </Box>
 );
 
 const SidebarFooter = () => (
-  <div className="sidebar-footer">
-    <Divider clearing />
-    <Image id="lattice-brand" src="https://tools.latticeautomation.com/seqviz/lattice-brand.png" />
-    <p>
+  <Box
+    sx={{
+      width: "100%",
+      textAlign: "center",
+      position: "absolute",
+      bottom: 0,
+      p: 1,
+    }}
+  >
+    <Divider sx={{ mb: 2 }} />
+    <img
+      src="https://tools.latticeautomation.com/seqviz/lattice-brand.png"
+      alt="Lattice Automation"
+      style={{ height: "64px", margin: "0 auto 8px auto" }}
+    />
+    <Typography variant="caption" display="block" sx={{ mb: 1 }}>
       Created by{" "}
-      <span>
-        <a href="https://latticeautomation.com/" rel="noopener noreferrer" target="_blank">
-          Lattice Automation
-        </a>
-      </span>
-    </p>
-    <p>
-      <Icon name="github" />
-      <span>
-        <a href="https://github.com/Lattice-Automation/seqviz" rel="noopener noreferrer" target="_blank">
+      <a
+        href="https://latticeautomation.com/"
+        rel="noopener noreferrer"
+        target="_blank"
+        style={{ fontWeight: "bold", color: "#1976d2" }}
+      >
+        Lattice Automation
+      </a>
+    </Typography>
+    <Typography variant="caption" display="block" sx={{ mb: 1 }}>
+      <Box component="span" sx={{ display: "inline-flex", alignItems: "center", gap: 0.5 }}>
+        <GitHubIcon sx={{ fontSize: "0.85rem" }} />
+        <a
+          href="https://github.com/Lattice-Automation/seqviz"
+          rel="noopener noreferrer"
+          target="_blank"
+          style={{ fontWeight: "bold", color: "#1976d2" }}
+        >
           seqviz
         </a>
-      </span>
-      <span>{"  |  "}</span>
-      <Icon name="medium" />
-      <span>
+      </Box>
+      {" | "}
+      <Box component="span" sx={{ display: "inline-flex", alignItems: "center", gap: 0.5 }}>
+        <ArticleIcon sx={{ fontSize: "0.85rem" }} />
         <a
           href="https://medium.com/@lattice.core/visualize-your-dna-sequences-with-seqviz-b1d945eb9684"
           rel="noopener noreferrer"
           target="_blank"
+          style={{ fontWeight: "bold", color: "#1976d2" }}
         >
           Story
         </a>
-      </span>
-      <span>{"  |  "}</span>
-      <Icon name="edit outline" />
-      <span>
+      </Box>
+      {" | "}
+      <Box component="span" sx={{ display: "inline-flex", alignItems: "center", gap: 0.5 }}>
+        <EditIcon sx={{ fontSize: "0.85rem" }} />
         <a
           href="https://docs.google.com/forms/d/1ILD3UwPvdkQlM06En7Pl9VqVpN_-g5iWs-B6gjKh9b0/viewform?edit_requested=true"
           rel="noopener noreferrer"
           target="_blank"
+          style={{ fontWeight: "bold", color: "#1976d2" }}
         >
           Survey
         </a>
-      </span>
-    </p>
-    <p>
-      <span>contact@latticeautomation.com</span>
-    </p>
-  </div>
+      </Box>
+    </Typography>
+    <Typography variant="caption" display="block">
+      contact@latticeautomation.com
+    </Typography>
+  </Box>
 );
+
