+++ demo/next.config.js
@@ -1,29 +1,30 @@
-const isProd = process.env.NODE_ENV === "production";
+const path = require('path');
 
-// const withTM = require("next-transpile-modules")(["seqviz", "../seqviz"]); // pass the modules you would like to see transpiled
-
-module.exports = {
-  // Use the CDN in production and localhost for development.
-  // https://nextjs.org/docs/api-reference/next.config.js/cdn-support-with-asset-prefix
-  assetPrefix: isProd ? "https://tools.latticeautomation.com/seqviz" : undefined,
-  // https://github.com/vercel/next.js/issues/33488#issuecomment-1063677133
-  images: isProd
-    ? {
-        domains: ["https://tools.latticeautomation.com/seqviz"],
-        path: "https://tools.latticeautomation.com/seqviz",
-      }
-    : undefined,
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  reactStrictMode: true,
+  output: 'export',
+  images: {
+    unoptimized: true,
+  },
   experimental: {
     externalDir: true,
   },
-  reactStrictMode: true,
   typescript: {
-    // !! WARN !!
-    // Dangerously allow production builds to successfully complete even if
-    // your project has type errors.
-    // !! WARN !!
-    // ../src/Circular/Annotations.tsx:179:9
-    // Type error: Type 'Properties<0 | (string & {}), string & {}>' is not assignable to type 'CSSProperties'.
+    // Allow builds to complete even with type errors from parent src directory
     ignoreBuildErrors: true,
   },
+  webpack: (config) => {
+    // Ensure only one instance of React is used
+    config.resolve.alias = {
+      ...config.resolve.alias,
+      react: path.resolve(__dirname, 'node_modules/react'),
+      'react-dom': path.resolve(__dirname, 'node_modules/react-dom'),
+    };
+    
+    return config;
+  },
 };
+
+module.exports = nextConfig;
+
