+++ demo/package.json
@@ -12,16 +12,20 @@
     "dev": "next dev -p 3010"
   },
   "dependencies": {
+    "@emotion/react": "^11.11.4",
+    "@emotion/server": "^11.11.0",
+    "@emotion/styled": "^11.11.5",
+    "@mui/icons-material": "^7.0.0",
+    "@mui/material": "^7.0.0",
+    "@mui/material-nextjs": "^7.0.0",
     "browserslist": "^4.21.3",
     "history": "^4.10.1",
     "lodash": "^4.17.21",
-    "next": "^13.5.6",
-    "next-with-less": "^2.0.5",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "semantic-ui-css": "^2.4.1",
-    "semantic-ui-react": "^2.1.3",
-    "typescript": "^4.8.2"
+    "next": "^15.5.5",
+    "react": "^19.2.0",
+    "react-dom": "^19.2.0",
+    "seqparse": "^0.2.1",
+    "typescript": "^5.5.0"
   },
   "browserslist": {
     "production": [
@@ -35,7 +39,11 @@
   },
   "devDependencies": {
     "@types/lodash": "^4.14.184",
-    "@types/react": "^18.0.18",
-    "webpack": "^5.76.0"
+    "@types/react": "^19.2.0",
+    "@types/react-dom": "^19.2.0",
+    "eslint": "^8.57.0",
+    "eslint-config-next": "^14.2.0",
+    "prettier": "^3.2.0",
+    "webpack": "^5.94.0"
   }
 }
