+++ demo/pages/_app.tsx
@@ -1,10 +1,26 @@
-import Head from "next/head";
 import * as React from "react";
-import "semantic-ui-css/semantic.min.css";
+import Head from "next/head";
+import { ThemeProvider, createTheme } from "@mui/material/styles";
+import CssBaseline from "@mui/material/CssBaseline";
+import type { AppProps } from "next/app";
 
 import "../styles/global.css";
 
-export default function App({ Component, pageProps }: any) {
+const theme = createTheme({
+  palette: {
+    primary: {
+      main: "#14b7db",
+    },
+    background: {
+      default: "#ffffff",
+    },
+  },
+  typography: {
+    fontFamily: "Roboto Mono, Monaco, monospace",
+  },
+});
+
+export default function App({ Component, pageProps }: AppProps) {
   return (
     <>
       <Head>
@@ -15,7 +31,11 @@ export default function App({ Component, pageProps }: any) {
         <meta name="viewport" content="width=device-width, initial-scale=1" />
         <meta name="theme-color" content="#000000" />
       </Head>
-      <Component {...pageProps} />
+      <ThemeProvider theme={theme}>
+        <CssBaseline />
+        <Component {...pageProps} />
+      </ThemeProvider>
     </>
   );
 }
+
