+++ demo/pages/index.tsx
@@ -3,3 +3,4 @@ import * as React from "react";
 import App from "../lib/App";
 
 export default () => <App />;
+
