--- demo2/.eslintrc.json
@@ -1,1 +1,0 @@
-{}
