--- demo2/README.md
@@ -1,16 +1,0 @@
-# seqviz Demo
-
-A demonstration of SeqViz. Visit [Lattice-Automation/seqviz](https://github.com/Lattice-Automation/seqviz) to see the source code and contribute.
-
-## Development
-
-To run the demo as part of development:
-
-```bash
-git clone https://github.com/Lattice-Automation/seqviz
-cd seqviz/demo
-npm install
-cd ..
-npm install
-npm start
-```
