--- demo2/pages/index.tsx
@@ -1,6 +1,0 @@
-import * as React from "react";
-
-import App from "../lib/App";
-
-export default () => <App />;
-
