--- demo2/public/favicon.ico
@@ -1,2 +1,0 @@
-# Placeholder for favicon - copy from demo folder if needed
-
